# Conflicts:
#	data/src/main/graphql/com/multimoney/multimoney/graphql/accountsmart/mutation/GlobalRequest.graphql
#	data/src/main/java/com/multimoney/data/networking/SmartAccountApi.kt
#	data/src/main/java/com/multimoney/data/repository/SmartAccountRepositoryImpl.kt

+++ app/src/main/java/com/multimoney/multimoney/presentation/ui/smart/SmartViewModel.kt
@@ -112,7 +112,7 @@ class SmartViewModel @Inject constructor(
             idAddressLevel2 = accountSmartData?.idAddressLevel2 ?: 0,
             idAddressLevel3 = accountSmartData?.idAddressLevel3 ?: 0,
             idEconomicActivity = accountSmartData?.idEconomicActivity ?: 0,
-            income = accountSmartData?.income?.toInt() ?: 0,
+            income = accountSmartData?.income?.toDouble() ?: 0.0,
             addressDetail = accountSmartData?.addressDetail ?: "",
             user = accountSmartData?.user ?: "",
             idBrand = accountSmartData?.idBrand ?: 0,
@@ -276,12 +276,14 @@ class SmartViewModel @Inject constructor(
             is OnContinueEnable -> uiState = uiState.copy(isContinueEnabled = event.enable)
             is OnLoadingValueChange -> uiState = uiState.copy(isLoading = event.isLoading)
             is OnOpenDialogValueChange -> uiState = uiState.copy(openDialog = event.openDialog)
-            is OnFailureWithDialog -> uiState =
-                uiState.copy(isLoading = event.isLoading, openDialog = event.openDialog)
+            is OnFailureWithDialog ->
+                uiState =
+                    uiState.copy(isLoading = event.isLoading, openDialog = event.openDialog)
             is OnNextStep -> nextStep()
             is OnPreviousStep -> previousStep()
-            is UIEvent.OnContinueVisible -> uiState =
-                uiState.copy(isContinueVisible = event.visible)
+            is UIEvent.OnContinueVisible ->
+                uiState =
+                    uiState.copy(isContinueVisible = event.visible)
             is OnCallMutationUpdateGlobalRequestUseCase -> onUpdateAccountSmartData(event.accountSmartData)
         }
     }
