+++ domain/src/main/java/com/multimoney/domain/interaction/accountsmart/MutationGlobalRequestUseCaseImpl.kt
@@ -21,7 +21,7 @@ class MutationGlobalRequestUseCaseImpl(val repository: SmartAccountRepository) :
         idAddressLevel2: Long,
         idAddressLevel3: Long,
         idEconomicActivity: Long,
-        income: Int,
+        income: Double,
         addressDetail: String,
         user: String,
         idBrand: Int,
